requirements.txt
pybit==5.8.0
pandas==2.2.3
numpy==2.2.3
python-dotenv==1.1.0
ta==0.11.0
schedule==1.2.2
requests==2.32.3
websockets==13.1
twscrape==0.17.0
